[OPLU-124][OPLU-130]: add icon button variant, align checkbox and radio error states and add missing focus to button (#11)
This PR covers the following:

Add icon button mode to the existing button component.
Verify that label and checkbox error states are aligned with designs, they are. 😄
Rename the focus in the theme to better reflect its purpose.
Add a style guide for the new component variant.
Add missing focus state for buttons.